Create all the elements and create the appender.
Created all elements, added relevant classes to them. This was done with
'.classList.toggle('class')'.

Set textContent to relevant arguments in Header func, append all of the
children to the header div.

Return new headerElement with it's new content.

Also set up headerAppender to work with a css selector, using
'document.querySelector(arg)', and simply appended the header to that
selected element.